Fix error message, make  exceptions

+++ api/app/Http/Controllers/Calendar/DeleteCalendarController.php
@@ -11,7 +11,12 @@ class DeleteCalendarController extends Controller
 {
     public function __invoke(Request $request): JsonResponse
     {
-        $calendar = Calendar::query()->where('slug', '=', $request->slug)->first();
+        $authenticatedUser = auth()->user();
+        $calendar = Calendar::query()->where('user_id', '=', $authenticatedUser->id)->where('slug', '=', $request->slug)->first();
+
+        if (!$calendar) {
+            return response()->json(['error' => 'Calendrier non trouvé'], 404);
+        }
         $calendar->delete();
         return response()->json(['message' => 'Votre calendrier à bien été supprimé']);
     }
